scripts/python/requirements.txt
boto3
psycopg2-binary
python-dotenv
pandas
pyarrow